requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn>=0.24.0
pyyaml>=6.0
pypdf2>=3.0
requests>=2.28.0
umap-learn>=0.5.0
numpy>=1.24.0
